Make a RGB version of etch-and-sketch
The webpage starts with a 16 x 16 grid. Hovering the mouse over any
square changes its color to a random one. Once the color is changed, it
cannot be changed again. The user has the option to change the number of
squares by clicking on the 'change size' button on the top. The button
will prompt the user to input an integer to make an n-by-n grid. The
code limits the number to 100.